LayoutTests/html5lib/generated/run-tests18-data-expected.txt

CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
../resources/tests18.dat: PASS
